requirements.txt
pandas
openpyxl
pillow
pyside6
